Revert "implement: usePortfolioFetch handle logic outside usePortfolio"
This reverts commit 29b3972fee5d5549543c24976446e8a85d19bca2.
